requirements.txt
mysql-connector
pytz
requests
lxml
